Refactor: Simplify SnapSyncStage implementation
Removes unused fields and methods, streamlines logic, and consolidates code.
