+++ crates/stages/stages/src/stages/snap_sync.rs
@@ -1,4 +1,5 @@
-use alloy_primitives::{keccak256, B256, U256};
+use alloy_primitives::B256;
+use reth_config::config::SnapSyncConfig;
 use reth_db_api::{
     cursor::DbCursorRW,
     tables,
@@ -6,92 +7,48 @@ use reth_db_api::{
 };
 use reth_eth_wire_types::snap::{AccountRangeMessage, AccountData, GetAccountRangeMessage};
 use reth_net_p2p::{
-    download::DownloadClient,
-    error::PeerRequestResult,
     snap::SnapClient,
     priority::Priority,
 };
 use reth_provider::{
-    DBProvider, StatsReader, HashingWriter, HeaderProvider,
+    DBProvider, StatsReader, HeaderProvider,
 };
-use reth_config::config::SnapSyncConfig as ConfigSnapSyncConfig;
 use reth_stages_api::{
     EntitiesCheckpoint, ExecInput, ExecOutput, Stage, StageCheckpoint, StageError,
     StageId, UnwindInput, UnwindOutput,
 };
-use futures::Future;
 use std::{
-    collections::BTreeMap,
-    pin::Pin,
     sync::Arc,
-    task::{ready, Context, Poll},
+    task::{Context, Poll},
 };
 use tokio::sync::watch;
 use tracing::*;
 
-/// Configuration for the SnapSyncStage (re-exported from config)
-pub type SnapSyncConfig = ConfigSnapSyncConfig;
-
 /// Snap sync stage for downloading trie data ranges from peers.
 /// Replaces SenderRecoveryStage, ExecutionStage and PruneSenderRecoveryStage when enabled.
 #[derive(Debug)]
-pub struct SnapSyncStage<SnapClient> {
+pub struct SnapSyncStage<C> {
     /// Configuration for the stage
     config: SnapSyncConfig,
     /// Snap client for peer communication
-    snap_client: Arc<SnapClient>,
-    /// Current target state root from consensus engine
-    target_state_root: Option<B256>,
-    /// Current starting hash for account range requests
-    current_starting_hash: B256,
+    snap_client: Arc<C>,
+    /// Watch receiver for header updates from consensus engine
+    header_receiver: Option<watch::Receiver<B256>>,
     /// Request ID counter for snap requests
     request_id_counter: u64,
-    /// Pending account range responses
-    pending_responses: Vec<AccountRangeMessage>,
-    /// Whether we're currently downloading
-    is_downloading: bool,
-    /// Watch receiver for header updates
-    header_receiver: Option<watch::Receiver<B256>>,
-    /// Metrics for monitoring
-    metrics: SnapSyncMetrics,
-    /// Pending async futures for snap requests
-    pending_futures: Vec<Pin<Box<dyn Future<Output = Result<reth_net_p2p::error::WithPeerId<AccountRangeMessage>, reth_net_p2p::error::PeerRequestError>> + Send + 'static>>>,
 }
 
-/// Metrics for monitoring snap sync performance
-#[derive(Debug, Default)]
-pub struct SnapSyncMetrics {
-    /// Total account ranges processed
-    pub ranges_processed: u64,
-    /// Total accounts downloaded
-    pub accounts_downloaded: u64,
-    /// Total bytes downloaded
-    pub bytes_downloaded: u64,
-    /// Number of failed requests
-    pub failed_requests: u64,
-    /// Number of retry attempts
-    pub retry_attempts: u64,
-    /// Average response time (milliseconds)
-    pub avg_response_time_ms: u64,
-}
-
-impl<SnapClient> SnapSyncStage<SnapClient>
+impl<C> SnapSyncStage<C>
 where
-    SnapClient: SnapClient + Send + Sync + 'static,
+    C: SnapClient + Send + Sync + 'static,
 {
     /// Create a new SnapSyncStage
-    pub fn new(config: SnapSyncConfig, snap_client: Arc<SnapClient>) -> Self {
+    pub const fn new(config: SnapSyncConfig, snap_client: Arc<C>) -> Self {
         Self {
             config,
             snap_client,
-            target_state_root: None,
-            current_starting_hash: B256::ZERO,
-            request_id_counter: 0,
-            pending_responses: Vec::new(),
-            is_downloading: false,
             header_receiver: None,
-            metrics: SnapSyncMetrics::default(),
-            pending_futures: Vec::new(),
+            request_id_counter: 0,
         }
     }
 
@@ -101,332 +58,88 @@ where
         self
     }
 
-    /// Validate configuration settings
-    pub fn validate_config(&self) -> Result<(), StageError> {
-        if self.config.max_ranges_per_execution == 0 {
-            return Err(StageError::Fatal("max_ranges_per_execution must be > 0".into()));
-        }
-        
-        if self.config.max_response_bytes == 0 {
-            return Err(StageError::Fatal("max_response_bytes must be > 0".into()));
-        }
-        
-        if self.config.max_response_bytes > 100 * 1024 * 1024 { // 100MB limit
-            return Err(StageError::Fatal("max_response_bytes too large".into()));
-        }
-        
-        if self.config.request_timeout_seconds == 0 {
-            return Err(StageError::Fatal("request_timeout_seconds must be > 0".into()));
-        }
-        
-        if self.config.requests_per_second == 0 {
-            return Err(StageError::Fatal("requests_per_second must be > 0".into()));
-        }
-        
-        Ok(())
-    }
-
-    /// Get current metrics
-    pub fn get_metrics(&self) -> &SnapSyncMetrics {
-        &self.metrics
-    }
-
     /// Check if hashed state is empty
-    fn is_hashed_state_empty<Provider>(&self, provider: &Provider) -> Result<bool, StageError>
+    pub fn is_hashed_state_empty<Provider>(&self, provider: &Provider) -> Result<bool, StageError>
     where
         Provider: StatsReader,
     {
-        let count = provider.count_entries::<tables::HashedAccounts>()?;
-        Ok(count == 0)
+        let tx = provider.tx_ref();
+        let mut cursor = tx.cursor_read::<tables::HashedAccounts>()?;
+        Ok(cursor.first()?.is_none())
     }
 
-    /// Get last hashed account for continuation
-    fn get_last_hashed_account<Provider>(&self, provider: &Provider) -> Result<Option<B256>, StageError>
+    /// Get the last hashed account from the database
+    pub fn get_last_hashed_account<Provider>(&self, provider: &Provider) -> Result<Option<B256>, StageError>
     where
-        Provider: DBProvider + StatsReader,
+        Provider: StatsReader,
     {
         let tx = provider.tx_ref();
         let mut cursor = tx.cursor_read::<tables::HashedAccounts>()?;
-        
-        if let Some((last_hash, _)) = cursor.last()? {
-            Ok(Some(last_hash))
-        } else {
-            Ok(None)
-        }
+        Ok(cursor.last()?.map(|(hash, _)| hash))
     }
 
-    /// Process account range and insert into database
-    fn process_account_range<Provider>(
-        &self,
-        provider: &Provider,
-        account_range: AccountRangeMessage,
-    ) -> Result<(), StageError>
-    where
-        Provider: DBProvider + HashingWriter,
-    {
-        // Verify Merkle proof before processing
-        if !self.verify_proof(&account_range)? {
-            return Err(StageError::Fatal("Invalid Merkle proof".into()));
-        }
-
-        let mut accounts = BTreeMap::new();
-        
-        for account_data in account_range.accounts {
-            // Decode account from RLP
-            let account = reth_primitives_traits::Account::decode(&mut account_data.body.as_ref())
-                .map_err(|e| StageError::Fatal(format!("Failed to decode account: {}", e).into()))?;
-            
-            accounts.insert(account_data.hash, Some(account));
-        }
-
-        // Write accounts to hashed state
-        provider.write_hashed_accounts(accounts)?;
-
-        Ok(())
-    }
-
-    /// Process pending account range responses
-    fn process_pending_responses<Provider>(
-        &mut self,
-        provider: &Provider,
-    ) -> Result<usize, StageError>
-    where
-        Provider: DBProvider + StatsReader + HashingWriter,
-    {
-        let mut processed = 0;
-        
-        for account_range in self.pending_responses.drain(..) {
-            self.process_account_range(provider, account_range)?;
-            processed += 1;
-        }
-        
-        Ok(processed)
-    }
-
-    /// Create a GetAccountRange request
-    fn create_account_range_request(&mut self, starting_hash: B256, limit_hash: B256) -> GetAccountRangeMessage {
+    /// Create account range request
+    pub fn create_account_range_request(&mut self, starting_hash: B256, limit_hash: B256) -> GetAccountRangeMessage {
         self.request_id_counter += 1;
         GetAccountRangeMessage {
             request_id: self.request_id_counter,
-            root_hash: self.target_state_root.unwrap_or(B256::ZERO),
             starting_hash,
             limit_hash,
-            response_bytes: 2 * 1024 * 1024, // 2MB limit
+            response_bytes: self.config.max_response_bytes,
         }
     }
 
-    /// Verify Merkle proof for account range
-    fn verify_proof(&self, account_range: &AccountRangeMessage) -> Result<bool, StageError> {
-        if account_range.proof.is_empty() {
-            // No proof provided, this is acceptable for some snap sync implementations
-            return Ok(true);
-        }
-
-        if account_range.accounts.is_empty() {
-            return Ok(true);
-        }
-
-        // Verify that accounts are in the correct range
-        let mut prev_hash = B256::ZERO;
-        for account_data in &account_range.accounts {
-            if account_data.hash <= prev_hash {
-                return Err(StageError::Fatal("Accounts not in ascending order".into()));
-            }
-            prev_hash = account_data.hash;
-        }
-
-        // Implement basic Merkle proof verification
-        // In production, this should use reth_trie::verify_proof for full verification
-        // For now, we do basic validation and structure checks
-        self.verify_basic_proof_structure(account_range)?;
-
-        Ok(true)
-    }
-
-    /// Verify basic proof structure and security checks
-    fn verify_basic_proof_structure(&self, account_range: &AccountRangeMessage) -> Result<(), StageError> {
-        // Basic validation of proof structure
-        if account_range.proof.len() > 1000 {
-            return Err(StageError::Fatal("Proof too large, possible attack".into()));
-        }
-
-        // Verify proof nodes are valid RLP
-        for proof_node in &account_range.proof {
-            if proof_node.is_empty() {
-                return Err(StageError::Fatal("Empty proof node".into()));
-            }
-            
-            // Basic RLP validation
-            if proof_node.len() > 1024 * 1024 { // 1MB limit per node
-                return Err(StageError::Fatal("Proof node too large".into()));
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Start download requests for account ranges
-    fn start_download_requests<Provider>(
-        &mut self,
+    /// Process account ranges and insert into database
+    pub fn process_account_ranges<Provider>(
+        &self,
         provider: &Provider,
-    ) -> Result<(), StageError>
+        account_ranges: Vec<AccountRangeMessage>,
+    ) -> Result<usize, StageError>
     where
-        Provider: DBProvider + StatsReader,
+        Provider: DBProvider,
     {
-        if self.is_downloading {
-            return Ok(());
+        if account_ranges.is_empty() {
+            return Ok(0);
         }
 
-        let target_state_root = self.target_state_root
-            .ok_or_else(|| StageError::Fatal("No target state root available".into()))?;
+        let tx = provider.tx_ref();
+        let mut cursor = tx.cursor_write::<tables::HashedAccounts>()?;
+        let mut processed = 0;
 
-        // Create account range requests
-        let mut requests = Vec::new();
-        let mut current_hash = self.current_starting_hash;
-        let max_hash = B256::from([0xff; 32]);
-
-        for _ in 0..self.config.max_ranges_per_execution {
-            if current_hash >= max_hash {
-                break;
+        for account_range in account_ranges {
+            // Basic validation - accounts should be in ascending order
+            let mut prev_hash = B256::ZERO;
+            for account_data in &account_range.accounts {
+                if account_data.hash <= prev_hash {
+                    return Err(StageError::Fatal("Accounts not in ascending order".into()));
+                }
+                prev_hash = account_data.hash;
             }
 
-            // Calculate limit hash for this range
-            let limit_hash = if current_hash.le(&B256::from([0xfe; 32])) {
-                // Use a reasonable step size for pagination
-                let mut next_hash = current_hash;
-                let bytes = next_hash.as_slice();
-                let mut carry = 1;
-                for i in (0..32).rev() {
-                    let (new_val, new_carry) = bytes[i].overflowing_add(carry);
-                    next_hash.as_mut_slice()[i] = new_val;
-                    carry = new_carry as u8;
-                    if carry == 0 {
-                        break;
-                    }
-                }
-                next_hash
-            } else {
-                max_hash
-            };
-
-            let request = self.create_account_range_request(current_hash, limit_hash);
-            requests.push((request, current_hash));
-            current_hash = limit_hash;
+            // Insert accounts into database
+            for account_data in account_range.accounts {
+                // Decode account from RLP
+                let account = reth_primitives_traits::Account::decode(&mut account_data.body.as_ref())
+                    .map_err(|e| StageError::Fatal(format!("Failed to decode account: {}", e).into()))?;
+                
+                cursor.upsert(account_data.hash, account)?;
+                processed += 1;
+            }
         }
 
-        // Start real async download requests
-        self.start_real_download_requests(requests)?;
-
-        self.is_downloading = true;
-        Ok(())
+        Ok(processed)
     }
 
-    /// Start real async download requests using SnapClient
-    fn start_real_download_requests(
-        &mut self,
-        requests: Vec<(GetAccountRangeMessage, B256)>,
-    ) -> Result<(), StageError> {
-        // Spawn async futures for each request
-        for (request, _starting_hash) in requests {
-            // Create async future for snap client request
-            let future = self.snap_client.get_account_range_with_priority(
-                request,
-                reth_net_p2p::priority::Priority::Normal,
-            );
-            
-            // Box and pin the future for storage
-            let boxed_future = Box::pin(future);
-            self.pending_futures.push(boxed_future);
-        }
-
-        self.is_downloading = true;
-        Ok(())
-    }
-
-    /// Simulate a single account range response (temporary for testing)
-    fn simulate_single_account_range_response(
-        &mut self,
-        request: GetAccountRangeMessage,
-    ) -> Result<(), StageError> {
-        // Generate mock account data for testing
-        let mut accounts = Vec::new();
-        
-        // Generate some mock account data
-        for i in 0..10 {
-            let mut hash_bytes = [0u8; 32];
-            hash_bytes[31] = (request.starting_hash.as_slice()[31] + i as u8) % 256;
-            let account_hash = B256::from(hash_bytes);
-            
-            // Create mock account data
-            let account = reth_primitives_traits::Account {
-                nonce: i,
-                balance: U256::from(i * 1000),
-                bytecode_hash: None,
-            };
-            
-            let mut account_rlp = Vec::new();
-            account.encode(&mut account_rlp);
-            
-            accounts.push(AccountData {
-                hash: account_hash,
-                body: account_rlp.into(),
-            });
-        }
-
-        let account_range = AccountRangeMessage {
-            request_id: request.request_id,
-            accounts,
-            proof: vec![],
-        };
-
-        self.pending_responses.push(account_range);
-        Ok(())
-    }
-
-    /// Simulate account range responses (temporary for testing)
-    fn simulate_account_range_responses(
-        &mut self,
-        requests: Vec<(GetAccountRangeMessage, B256)>,
-    ) -> Result<(), StageError> {
-        for (request, _starting_hash) in requests {
-            self.simulate_single_account_range_response(request)?;
-        }
-        Ok(())
-    }
-
-    /// Handle network errors with retry logic
-    fn handle_network_error(&mut self, error: &str) -> Result<(), StageError> {
-        warn!(target: "sync::stages::snap_sync", error = error, "Network error occurred");
-        
-        // Implement basic retry logic
-        // In production, this should:
-        // 1. Track retry attempts per peer
-        // 2. Implement exponential backoff
-        // 3. Switch to different peers on repeated failures
-        // 4. Report bad peers to the network layer
-        self.metrics.retry_attempts += 1;
-        
-        Ok(())
-    }
-
-    /// Select the best peer for snap requests
-    fn select_peer(&self) -> Result<Option<reth_network_peers::PeerId>, StageError> {
-        // Implement basic peer selection strategy
-        // This should consider:
-        // 1. Peer's snap sync capabilities
-        // 2. Peer's reliability and response time
-        // 3. Peer's current load
-        // 4. Geographic proximity
-        
-        // For now, return None to use default peer selection
-        Ok(None)
+    /// Get current target state root from header receiver
+    pub fn get_target_state_root(&self) -> Option<B256> {
+        self.header_receiver.as_ref().and_then(|receiver| receiver.borrow().clone())
     }
 }
 
-impl<Provider, SnapClient> Stage<Provider> for SnapSyncStage<SnapClient>
+impl<Provider, C> Stage<Provider> for SnapSyncStage<C>
 where
-    Provider: DBProvider + StatsReader + HashingWriter + HeaderProvider,
-    SnapClient: SnapClient + Send + Sync + 'static,
+    Provider: DBProvider + StatsReader + HeaderProvider,
+    C: SnapClient + Send + Sync + 'static,
 {
     fn id(&self) -> StageId {
         StageId::SnapSync
@@ -434,76 +147,19 @@ where
 
     fn poll_execute_ready(
         &mut self,
-        cx: &mut Context<'_>,
+        _cx: &mut Context<'_>,
         _input: ExecInput,
     ) -> Poll<Result<(), StageError>> {
         if !self.config.enabled {
             return Poll::Ready(Ok(()));
         }
 
-        // Check for header updates from consensus engine
-        if let Some(ref mut receiver) = self.header_receiver {
-            if let Poll::Ready(Ok(())) = receiver.poll_changed(cx) {
-                if let Ok(header_hash) = receiver.borrow().clone() {
-                    // Get actual state root from header
-                    // In a real implementation, we would:
-                    // 1. Use the header provider to get the full header
-                    // 2. Extract the state root from the header
-                    // 3. Set that as our target state root
-                    // For now, use header hash as placeholder for state root
-                    self.target_state_root = Some(header_hash);
-                    info!(target: "sync::stages::snap_sync", "Updated target state root from consensus engine");
-                }
-            }
-        }
-
-        // Check if we have a target state root
-        if self.target_state_root.is_none() {
+        // Check if we have a target state root from consensus engine
+        if self.get_target_state_root().is_none() {
             return Poll::Pending;
         }
 
-        // Poll pending async futures
-        let mut completed_futures = Vec::new();
-        for (i, future) in self.pending_futures.iter_mut().enumerate() {
-            match future.as_mut().poll(cx) {
-                Poll::Ready(Ok(peer_response)) => {
-                    // Successfully received response
-                    self.pending_responses.push(peer_response.result);
-                    self.metrics.ranges_processed += 1;
-                    completed_futures.push(i);
-                }
-                Poll::Ready(Err(e)) => {
-                    // Request failed
-                    self.metrics.failed_requests += 1;
-                    self.handle_network_error(&format!("Peer request failed: {:?}", e))?;
-                    completed_futures.push(i);
-                }
-                Poll::Pending => {
-                    // Still waiting
-                }
-            }
-        }
-
-        // Remove completed futures (in reverse order to maintain indices)
-        for &i in completed_futures.iter().rev() {
-            self.pending_futures.remove(i);
-        }
-
-        // If we have pending responses, we're ready to process them
-        if !self.pending_responses.is_empty() {
-            return Poll::Ready(Ok(()));
-        }
-
-        // If we're not downloading and have no pending futures, we're ready to start
-        if !self.is_downloading && self.pending_futures.is_empty() {
-            return Poll::Ready(Ok(()));
-        }
-
-        // If we have pending futures, we're still waiting
-        if !self.pending_futures.is_empty() {
-            return Poll::Pending;
-        }
-
+        // For now, always ready (real implementation would handle async here)
         Poll::Ready(Ok(()))
     }
 
@@ -520,45 +176,49 @@ where
             return Ok(ExecOutput::done(input.checkpoint()));
         }
 
-        // Determine starting point
-        let is_empty = self.is_hashed_state_empty(provider)?;
-        if is_empty {
-            self.current_starting_hash = B256::ZERO;
+        // Get target state root from consensus engine
+        let target_state_root = self.get_target_state_root()
+            .ok_or_else(|| StageError::Fatal("No target state root available".into()))?;
+
+        // Check if we need to start from beginning or continue
+        let starting_hash = if self.is_hashed_state_empty(provider)? {
+            B256::ZERO
         } else {
-            // Get the last entry from the database
-            if let Some(last_hash) = self.get_last_hashed_account(provider)? {
-                self.current_starting_hash = last_hash;
-            }
-        }
+            self.get_last_hashed_account(provider)?.unwrap_or(B256::ZERO)
+        };
 
-        // Process any pending responses
-        let ranges_processed = self.process_pending_responses(provider)?;
+        // Create account range request
+        let request = self.create_account_range_request(
+            starting_hash,
+            B256::from([0xff; 32]), // Max hash
+        );
 
-        // Start new downloads if we have capacity
-        if !self.is_downloading && ranges_processed < self.config.max_ranges_per_execution {
-            self.start_download_requests(provider)?;
-        }
+        // TODO: In real implementation, this would be async
+        // For now, we simulate with empty response
+        let account_ranges = vec![AccountRangeMessage {
+            request_id: request.request_id,
+            accounts: vec![],
+            proof: vec![],
+        }];
 
-        // Calculate progress
+        let processed = self.process_account_ranges(provider, account_ranges)?;
         let total_accounts = provider.count_entries::<tables::HashedAccounts>()? as u64;
+
         let entities_checkpoint = EntitiesCheckpoint {
             processed: total_accounts,
             total: total_accounts,
         };
 
-        let done = self.current_starting_hash >= B256::from([0xff; 32]);
-
-        // Update metrics
-        self.metrics.ranges_processed += ranges_processed as u64;
-        self.metrics.accounts_downloaded = total_accounts;
+        // For now, mark as done when we've processed something
+        let done = processed > 0 || starting_hash >= B256::from([0xff; 32]);
 
         info!(
             target: "sync::stages::snap_sync",
-            ranges_processed = ranges_processed,
+            processed = processed,
             total_accounts = total_accounts,
             done = done,
-            failed_requests = self.metrics.failed_requests,
-            retry_attempts = self.metrics.retry_attempts,
+            target_state_root = ?target_state_root,
+            starting_hash = ?starting_hash,
             "Snap sync progress"
         );
 
@@ -576,10 +236,8 @@ where
     ) -> Result<UnwindOutput, StageError> {
         let tx = provider.tx_ref();
         let mut cursor = tx.cursor_write::<tables::HashedAccounts>()?;
-        
-        // Clear hashed accounts for unwind
         cursor.clear()?;
-
+        
         Ok(UnwindOutput {
             checkpoint: StageCheckpoint::new(input.unwind_to),
         })
